requirements.txt
streamlit==1.39.0
python-docx==1.1.2
